perf(catalog): 7-49x faster find + 8 bytes/entry (Phase 1, non-breaking)
S3 - route the CLI find path (FindService) through the synchronous FindOptions.Find
instead of the work-stealing FindAsync. The async path ran every entry through an
async Task<bool> state machine plus Task.Yield/Task.Delay every 500/2000 entries.
Measured on a 1M-entry catalog: name search 1746ms -> 36ms (48x), path search
1755ms -> 236ms (7.4x), regex name 1817ms -> 60ms (30x). cdeWin already used the
sync path; only the CLI was on the slow one.

R1a - FileEntryCount/DirEntryCount changed long -> uint on DirEntry/RootEntry/
ICommonEntry. They were already (uint)-truncated at assignment, so semantics are
unchanged. In-memory only, no catalog format impact. Confirmed 211.94 -> 203.94 B/entry.

Tests: cdeLibTest 126 passed. Format-affecting work (path-alloc S1, file/dir split,
hash side-table) is deferred to Phases 2-3.

+++ src/cdeLib/Entities/ICommonEntry.cs
@@ -33,8 +33,8 @@ public interface ICommonEntry
 
     bool IsDirectory { get; set; }
     bool PathProblem { get; }
-    long FileEntryCount { get; set; }
-    long DirEntryCount { get; set; }
+    uint FileEntryCount { get; set; }
+    uint DirEntryCount { get; set; }
     DateTime Modified { get; set; }
     bool IsHashDone { get; set; }
     bool IsPartialHash { get; set; }
